add tests for deposed nodes (both plan and apply) + rename deposed file because deposed can be destroyed and forgotten
